Restructure API into routes
